python_testing/coding-test-testing-fromDisco/requirements.txt
black==22.10.0
psycopg2-binary==2.9.5
pytest==7.2.0
python-dotenv==0.21.0